+++ server/component/style/StyleModel.php
@@ -347,17 +347,27 @@ class StyleModel extends BaseModel implements IStyleModel
      * the form id of the form that we want to fetcht
      * @param int $record_id
      * the record id of the form entry
+     * @param int $own_entries_only
+     * If true it loads only records created by the same user
      * @retval array
      * the result of the fetched form row
      */
-    protected function fetch_entry_record($form_id, $record_id)
+    protected function fetch_entry_record($form_id, $record_id, $own_entries_only = 1)
     {
         $filter = " AND deleted = 0 AND record_id = " . $record_id;
-        $sql = 'CALL get_form_data_with_filter(:form_id, :filter)';
-        return $this->db->query_db_first($sql, array(
+        $sql = 'CALL get_form_data_for_user_with_filter(:form_id, :user_id, "'.$filter.'")';
+        $params = array(
             ":form_id" => $form_id,
-            ":filter" => $filter
-        ));
+            ":user_id" => $_SESSION['id_user']
+        );
+        if (!$own_entries_only) {
+            $sql = 'CALL get_form_data_with_filter(:form_id, "'.$filter.'")';
+            $params = array(
+                ":form_id" => $form_id
+            );
+        }
+        
+        return $this->db->query_db_first($sql, $params);
     }
 
     /* Public Methods *********************************************************/