+++ server/component/style/input/InputView.php
@@ -126,7 +126,6 @@ class InputView extends FormFieldView
         else if($this->value === null)
             $this->value = $this->default_value;
         if($this->entry_data){
-            $orig_val = $this->model->get_db_field("value", "");
             $this->value = $this->get_entry_value($this->entry_data, $this->value); 
         }
         if(
@@ -142,8 +141,6 @@ class InputView extends FormFieldView
     public function output_content_mobile()
     {        
         $style = parent::output_content_mobile();
-        // $curr_value = $this->model->get_form_field_value();
-        // $style['value']['content'] = $curr_value  ? $curr_value : ($this->type == "checkbox" ? $curr_value : $this->default_value);
         $style['value']['content'] = $this->value;
         $style['value']['default'] = $this->default_value;
         return $style;