+++ server/db/update_v4.0.0_v4.1.0.sql
@@ -67,12 +67,23 @@ INSERT INTO `fields` (`id`, `name`, `id_type`, `display`) VALUES (NULL, 'icon', 
 INSERT INTO `styles_fields` (`id_styles`, `id_fields`, `default_value`, `help`) 
 VALUES (get_style_id('tab'), get_field_id('icon'), '', 'Show icon; For web font awsome icons are used; For mobile ionicicons are used.');
 
-# executed on studybuddy
-
 INSERT INTO `fields` (`id`, `name`, `id_type`, `display`) VALUES (NULL, 'email_address', get_field_type_id('text'), '0');
 INSERT INTO `styles_fields` (`id_styles`, `id_fields`, `default_value`, `help`) 
 VALUES (get_style_id('formUserInput'), get_field_id('email_address'), '@email_user', 'Use `@email_user` to retrive automaticaly the user email. Emails are separated by the MAIL_SEPARATOR. It is `;`');
 
+# executed on studybuddy
 
+-- Add new field `own_entries_only` in style entryList
+INSERT INTO `fields` (`id`, `name`, `id_type`, `display`) VALUES (NULL, 'own_entries_only', get_field_type_id('checkbox'), '0');
+INSERT INTO `styles_fields` (`id_styles`, `id_fields`, `default_value`, `help`) 
+VALUES (get_style_id('entryList'), get_field_id('own_entries_only'), '1', 'If selected the entry list will load only the records entered by the user.');
+
+-- Add new field `own_entries_only` in style entryRecord
+INSERT INTO `styles_fields` (`id_styles`, `id_fields`, `default_value`, `help`) 
+VALUES (get_style_id('entryRecord'), get_field_id('own_entries_only'), '1', 'If selected the entry list will load only the records entered by the user.');
+
+-- Add new field `own_entries_only` in style entryRecord
+INSERT INTO `styles_fields` (`id_styles`, `id_fields`, `default_value`, `help`) 
+VALUES (get_style_id('formUserInput'), get_field_id('own_entries_only'), '1', 'If selected the entry list will load only the records entered by the user.');
 
 #gen all functions and procs