+++ CHANGELOG.md
@@ -23,8 +23,10 @@
     - `alert`
     - `link`
     - `rawtext`
+    - `textarea`
  - load all custom css files in the mobile app
  - set locale from mobile calls
+ - add field `own_entries_only` in style `formUserInput`. 
 
 ### Changes
  - `formUserInput` can be added to a page with path containing `record_id`. Example: `/courseViewAdvanced/[i:record_id]?`. When the id is loaded the entry will be in edit mode. Also an option for delete will apear.